move generation files into elk + change to default_config
